move lidccd package to image.lidccd
