+++ digicampipe/image/lidccd/kernels.py
