+++ digicampipe/image/lidccd/nova_client.py
