+++ digicampipe/image/lidccd/utils.py
