+++ pointing_from_stars.py
@@ -1,9 +1,9 @@
-from pkg_resources import resource_filename
 from glob import glob
 
 from digicampipe.image.sky_image import LidCCDObservation
-from digicampipe.image.utils import Rectangle
+from pkg_resources import resource_filename
 
+from digicampipe.image.lidccd.utils import Rectangle
 
 example_lid_CCD_image_file_paths = glob(
     resource_filename('digicampipe', 'tests/resources/stars_on_lid/*.fits')
